requirements.txt
Flask==2.3.2
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.6
MarkupSafe==2.1.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2